CRUD Funcionando

+++ SiADi/Modelo/Persona.cs
@@ -11,13 +11,13 @@ namespace SiADi.Modelo
     public class Persona : Base
     {
         public int Dni { set; get; }
-        public int Cuil { set; get; }
+        public long Cuil { set; get; }
         public string Nombre { set; get; }
         public string Apellido { set; get; }
         public DateTime Fecha_nacimiento { set; get; }
         public int Edad { set; get; }
         public string Direccion { set; get; }
-        public int Telefono { set; get; }
+        public long Telefono { set; get; }
         public string Codigo_qr { set; get; }
         public string Foto { set; get; }
         public Cargo Cargo { set; get; }
